+++ bugsnag-android-core/src/test/java/com/bugsnag/android/DeliveryDelegateTest.kt
@@ -111,7 +111,7 @@ internal class DeliveryDelegateTest {
         assertEquals(DeliveryStatus.DELIVERED, status)
         assertEquals("Sent 1 new event to Bugsnag", logger.msg)
 
-        val breadcrumb = requireNotNull(breadcrumbState.store.peek())
+        val breadcrumb = requireNotNull(breadcrumbState.copy().first())
         assertEquals(BreadcrumbType.ERROR, breadcrumb.type)
         assertEquals("java.lang.RuntimeException", breadcrumb.message)
         assertEquals("java.lang.RuntimeException", breadcrumb.metadata!!["errorClass"])