+++ CHANGELOG.md
@@ -38,6 +38,9 @@
 * Cache results from PackageManager
   [#1288](https://github.com/bugsnag/bugsnag-android/pull/1288)
 
+* Use ring buffer to store breadcrumbs
+  [#1286](https://github.com/bugsnag/bugsnag-android/pull/1286)
+
 ## 5.9.4 (2021-05-26)
 
 * Unity: add methods for setting autoNotify and autoDetectAnrs